nginx: do not reduce max processes
Use the upstream value for NGX_MAX_PROCESSES (=1024) again.
So it is safe to use `worker_processes auto;` (which sets it to
the number of CPUs). Also use this directive in the conf instead
of calculating the number in the init and passing it as argument.

Signed-off-by: Peter Stadler <peter.stadler@student.uibk.ac.at>
(cherry picked from commit 62609f1587b2e272dc8a2245e36ca6dc7d0c8820)